maintenance: prepare for ocamlformat
There is no Dune on Havana, so call ocamlformat from the Makefile
with in-place formatting instead.

This should make it easier to backport changes from Stockholm/trunk.

The .ocamlformat config is the same as the one on master.

ocamlformat isn't available in xs-opam for Havana either,
but we can use the one from trunk
(either installed through opam or in a Koji buildenv).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>