CA-351685: Reduce number of trace entries to 16 to reduce memory usage
This is part of an alternative fix for XSA-354.
Instead of multiplying memory usage 128x, we just multiply 16x
(in practice 4x due to request/reply for VM.stat taking up 3 small
 entries).

The trace buffer is used for debugging purposed only, and if needed
we can now increase the size of the trace buffer using the configuration
file.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>