CP-37368: use dune configurator and put sbin in correct place
Previously we supported an override only for sbindir, but that is not
supported by dune install.
In dune install we can override --prefix, --libdir, --etcdir, --docdir.
Binaries always go to prefix/bin and prefix/sbin.

So instead allow overriding the prefix, keeping only the OCaml libraries
under the opam root, and installing everything else to system
destinations when building from a specfile.

When built from opam and calling configure without arguments we set up
prefix and libdir such that it ends up where opam expects them.
This relies on findlib.

The command-line interface of Dune's configurator is not as advanced as
cmdline (uses built-in Arg), but suffices for our purpose here.

Also simplifies the wrapper configure script which no longer needs to
set up a temporary build dir: we call dune directly to run the
configurator.

If needed this can also generate OCaml files by adding appropriate rules
in the `dune` file. This is not done for `config.mk` which needs to stay
at the root for `make`.

DESTDIR is an install time variable, not a configure-time variable, so
reflect this in the Makefile (changing DESTDIR doesn't require
rebuilding).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>