CP-37931: Unit tests: do not require systemd connection
Unit tests run inside rpmbuild/mock may not have access to systemd.
Previously this would've failed with an exception (which is why haven't
been running the unit test at all in the .spec file).

Detect when we are not running under systemd and avoid raising an
exception, but allow exceptions to propagate when we run under systemd.

Checking whether we run under systemd or not may raise an exception
itself, so ignore that and treat it as not running under systemd.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>